deduplicate imports
Resolving merge conflicts in imports list formatted by simformat was
annoying because the changes are word-based, not line-based. With this
change, we can just delete the conflict markers and run simformat on
lines containing both conflicting versions, and the duplicate imports
will be merged into one. We can then delete the imports which GHC says
aren't used, that's much easier.